Renamed component and turned into QuestionCard
